refactor: 에러 파일 패키지 분리
