+++ src/main/java/com/devnity/devnity/domain/admin/service/AdminCourseService.java
@@ -1,8 +1,7 @@
 package com.devnity.devnity.domain.admin.service;
 
-import com.devnity.devnity.common.error.exception.BusinessException;
-import com.devnity.devnity.common.error.exception.EntityNotFoundException;
-import com.devnity.devnity.common.error.exception.ErrorCode;
+import com.devnity.devnity.web.error.exception.EntityNotFoundException;
+import com.devnity.devnity.web.error.exception.ErrorCode;
 import com.devnity.devnity.domain.admin.dto.request.CourseRequest;
 import com.devnity.devnity.domain.admin.dto.response.CourseResponse;
 import com.devnity.devnity.domain.user.entity.Course;
