+++ src/main/java/com/devnity/devnity/domain/admin/service/AdminRetrieveService.java
@@ -1,10 +1,9 @@
 package com.devnity.devnity.domain.admin.service;
 
-import com.devnity.devnity.common.error.exception.EntityNotFoundException;
-import com.devnity.devnity.common.error.exception.ErrorCode;
+import com.devnity.devnity.web.error.exception.EntityNotFoundException;
+import com.devnity.devnity.web.error.exception.ErrorCode;
 import com.devnity.devnity.domain.admin.entity.Invitation;
 import com.devnity.devnity.domain.admin.repository.InvitationRepository;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
