+++ src/main/java/com/devnity/devnity/domain/gather/service/GatherRetrieveService.java
@@ -1,7 +1,7 @@
 package com.devnity.devnity.domain.gather.service;
 
-import com.devnity.devnity.common.error.exception.EntityNotFoundException;
-import com.devnity.devnity.common.error.exception.ErrorCode;
+import com.devnity.devnity.web.error.exception.EntityNotFoundException;
+import com.devnity.devnity.web.error.exception.ErrorCode;
 import com.devnity.devnity.domain.gather.repository.GatherApplicantRepository;
 import com.devnity.devnity.domain.gather.repository.GatherCommentRepository;
 import com.devnity.devnity.domain.gather.repository.GatherRepository;
