+++ src/main/java/com/devnity/devnity/domain/introduction/entity/IntroductionComment.java
@@ -1,9 +1,9 @@
 package com.devnity.devnity.domain.introduction.entity;
 
-import static com.devnity.devnity.common.error.exception.ErrorCode.INTRODUCTION_COMMENT_ALREADY_DELETED;
-import static com.devnity.devnity.common.error.exception.ErrorCode.INTRODUCTION_COMMENT_IS_CHILD;
+import static com.devnity.devnity.web.error.exception.ErrorCode.INTRODUCTION_COMMENT_ALREADY_DELETED;
+import static com.devnity.devnity.web.error.exception.ErrorCode.INTRODUCTION_COMMENT_IS_CHILD;
 
-import com.devnity.devnity.common.error.exception.InvalidValueException;
+import com.devnity.devnity.web.error.exception.InvalidValueException;
 import com.devnity.devnity.domain.base.BaseEntity;
 import com.devnity.devnity.domain.user.entity.User;
 import javax.persistence.Column;
