+++ src/main/java/com/devnity/devnity/domain/introduction/service/IntroductionLikeService.java
@@ -1,10 +1,10 @@
 package com.devnity.devnity.domain.introduction.service;
 
-import static com.devnity.devnity.common.error.exception.ErrorCode.INTRODUCTION_LIKE_DUPLICATE;
-import static com.devnity.devnity.common.error.exception.ErrorCode.INTRODUCTION_LIKE_NOT_FOUND;
+import static com.devnity.devnity.web.error.exception.ErrorCode.INTRODUCTION_LIKE_DUPLICATE;
+import static com.devnity.devnity.web.error.exception.ErrorCode.INTRODUCTION_LIKE_NOT_FOUND;
 
-import com.devnity.devnity.common.error.exception.EntityNotFoundException;
-import com.devnity.devnity.common.error.exception.InvalidValueException;
+import com.devnity.devnity.web.error.exception.EntityNotFoundException;
+import com.devnity.devnity.web.error.exception.InvalidValueException;
 import com.devnity.devnity.domain.introduction.entity.IntroductionLike;
 import com.devnity.devnity.domain.introduction.respository.IntroductionLikeRepository;
 import lombok.RequiredArgsConstructor;
