+++ src/main/java/com/devnity/devnity/web/error/JwtAccessDeniedHandler.java
@@ -1,6 +1,6 @@
-package com.devnity.devnity.common.error;
+package com.devnity.devnity.web.error;
 
-import com.devnity.devnity.common.error.exception.AuthErrorCode;
+import com.devnity.devnity.web.error.exception.AuthErrorCode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import javax.servlet.ServletException;
