+++ src/main/java/com/devnity/devnity/web/error/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package com.devnity.devnity.common.error.exception;
+package com.devnity.devnity.web.error.exception;
 
 public class BusinessException extends RuntimeException {
