+++ src/main/java/com/devnity/devnity/web/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.devnity.devnity.web.user.service;
 
-import com.devnity.devnity.common.error.exception.ErrorCode;
-import com.devnity.devnity.common.error.exception.InvalidValueException;
+import com.devnity.devnity.web.error.exception.ErrorCode;
+import com.devnity.devnity.web.error.exception.InvalidValueException;
 import com.devnity.devnity.domain.gather.dto.SimpleGatherInfoDto;
 import com.devnity.devnity.domain.gather.service.GatherRetrieveService;
 import com.devnity.devnity.domain.mapgakco.dto.SimpleMapgakcoInfoDto;
