+++ src/test/java/com/devnity/devnity/domain/introduction/entity/IntroductionCommentTest.java
@@ -3,7 +3,7 @@ package com.devnity.devnity.domain.introduction.entity;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.devnity.devnity.common.error.exception.InvalidValueException;
+import com.devnity.devnity.web.error.exception.InvalidValueException;
 import com.devnity.devnity.domain.user.entity.User;
 import com.devnity.devnity.domain.user.entity.UserRole;
 import org.junit.jupiter.api.DisplayName;
